% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2Re.R
\name{chi2Re}
\alias{chi2Re}
\title{Il Chiquadro relativo}
\usage{
chi2Re(valIn1, valIn2)
}
\arguments{
\item{valIn1}{collezione che contine dati della prima variabile.}

\item{valIn2}{collezione che contiene dati della seconda variabile.}
}
\value{
il chiquadro relativo
}
\description{
Calcola il valore del chiquadro relativo per una coppia di collezioni
di osservazioni.
}
\examples{
\dontrun{
testVal1 <-  rep(c(1,2),c(5,5))
testVal2 <-  rep(c(1,2),c(5,5))
chi2Re(testVal1,testVal2)

testVal1 <-   c(1,1,2,2,1,1,2,2)
testVal2 <-  c(1,2,1,2,1,2,1,2)
chi2Re(testVal1,testVal2)
}

}
