% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffVariaPerCen.R
\name{coeffVariaPerCen}
\alias{coeffVariaPerCen}
\title{Il coefficiente di variazione percentuale}
\usage{
coeffVariaPerCen(collezione)
}
\arguments{
\item{collezione}{la collezione dei dati da analizzare.}
}
\value{
coefficiente di variazione percentuale
}
\description{
La funzione calcola il valore del coefficiente di
variazione percentuale.
}
\examples{

coeffVariaPerCen(c(3,1,2,3,52,NA,NA,23))

}
