% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFATR.R
\docType{data}
\name{DFATR}
\alias{DFATR}
\title{Concentrazione di antociani per giorno}
\format{
Vedere il manuale del corso.
}
\usage{
DFATR
}
\description{
Un dataset che contiene le concentrazioni di antociani per tank-giorno.
}
\keyword{datasets}
