% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeMedSemp.R
\name{diffeMedSemp}
\alias{diffeMedSemp}
\title{La differenza semplice media}
\usage{
diffeMedSemp(valoriVaria)
}
\arguments{
\item{valoriVaria}{la collezione che contine dati della  variabile.}
}
\value{
la differenza semplice media
}
\description{
Calcola il valore della statistica differenza semplice media.
}
\examples{

testVal1 <-   c(1,1,2,2,1,1,2,2)
diffeMedSemp(testVal1)

testVal2 <-   c(1,1,2,NA,1,1,2,2)
diffeMedSemp(testVal2)


}
