% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espeDue.R
\docType{data}
\name{espeDue}
\alias{espeDue}
\title{Barattoli di conserva}
\format{
Vedere il manuale del corso.
}
\usage{
espeDue
}
\description{
Un dataset che contiene le caratteristiche  di barattoli
prodotti in una azienda.
}
\keyword{datasets}
