% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esperimentoUno.R
\docType{data}
\name{esperimentoUno}
\alias{esperimentoUno}
\title{Barattoli di conserva (dati completi)}
\format{
Vedere il manuale del corso.
}
\usage{
esperimentoUno
}
\description{
Un dataset che contiene le caratteristiche  di barattoli
prodotti in una azienda, senza missing values.
}
\keyword{datasets}
