% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishDF.R
\docType{data}
\name{fishDF}
\alias{fishDF}
\title{Caso di studio fishcatch}
\format{
An object of class \code{data.frame} with 159 rows and 11 columns.
}
\usage{
data(fishDF, package="FMStatR")

fishDF
}
\description{
Un dataset che contiene il caso di studio fishcatch.
}
\references{
\url{http://jse.amstat.org/datasets/fishcatch.txt}

Brofeldt, Pekka: Bidrag till kaennedom on fiskbestondet i vaara
sjoear. Laengelmaevesi. T.H.Jaervi: Finlands Fiskeriet  Band 4,
Meddelanden utgivna av fiskerifoereningen i Finland.
Helsingfors 1917
}
\keyword{datasets}
