% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris2.R
\docType{data}
\name{iris2}
\alias{iris2}
\title{Caso di studio iris2}
\format{Vedere il manuale del corso.}
\usage{
iris2
}
\description{
Un dataset che contiene il caso di studio iris con dati
trasformati.
}
\keyword{datasets}
