% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris7.R
\docType{data}
\name{iris7}
\alias{iris7}
\title{Caso di studio iris7}
\format{Vedere il manuale del corso.}
\usage{
iris7
}
\description{
Un dataset che contiene il caso di studio iris con dati
trasformati.
}
\keyword{datasets}
