% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meloniDF.R
\docType{data}
\name{meloniDF}
\alias{meloniDF}
\title{Caso di studio meloniDF}
\format{Vedere il manuale del corso.
\describe{
\item{Altezza}{ l'altezza}
\item{Diametro}{ il diametro}
\item{PresenzaPatogeno}{ indicatore binario}
}}
\usage{
meloniDF
}
\description{
Un dataset che contiene il caso di studio meloniDF.
}
\keyword{datasets}
