% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\docType{data}
\name{rats}
\alias{rats}
\title{Caso di studio rats, modificato}
\format{
Vedere il manuale del corso.
\describe{
\item{time}{ minuti prima di osservare la morte dell'animale}
\item{poison}{ tipo di veleno}
\item{integra}{ tipo di integratore alimentare}
}
}
\source{
\url{https://cran.r-project.org/doc/contrib/Faraway-PRA.pdf}
}
\usage{
data(rats, package="FMStatR")

rats
}
\description{
Un dataset che contiene il caso di studio rats con nome
variabili modificate.
}
\references{
Box G and Cox D (1964) "An analysis of transformations" 
J. Roy. Stat. Soc. Series B.26211.

\url{https://cran.r-project.org/web/packages/faraway/index.html}
}
\keyword{datasets}
