% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riceDF.R
\docType{data}
\name{riceDF}
\alias{riceDF}
\title{Caso di studio riceDF}
\format{
An object of class \code{data.frame} with 72 rows and 7 columns.
}
\usage{
data(riceDF, package="FMStatR")

riceDF
}
\description{
Un dataset che contiene il caso di studio riceDF.
}
\references{
\url{https://cran.r-project.org/web/packages/DAAG/index.html}

Perrine, F.M., Prayitno, J., Weinman, J.J., Dazzo, F.B. and 
Rolfe, B. 2001. Rhizobium plasmids are involved in the inhibition or 
stimulation of rice growth and development.
 Australian Journal of Plant Physiology 28: 923-927.
}
\keyword{datasets}
