% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serieTempera.R
\docType{data}
\name{serieTempera}
\alias{serieTempera}
\title{Caso di studio serieTempera}
\format{Vedere il manuale del corso.}
\usage{
serieTempera
}
\description{
Una lista che contiene i dati dello studio.
}
\keyword{datasets}
