% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spigmais.R
\docType{data}
\name{spigmais}
\alias{spigmais}
\title{Caso di studio spigmais}
\format{Vedere il manuale del corso.}
\usage{
spigmais
}
\description{
Un dataset che contiene il caso di studio spigmais
}
\keyword{datasets}
