% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valoriModali.R
\name{valoriModali}
\alias{valoriModali}
\title{La moda e i valori modali}
\usage{
valoriModali(collezioneValori)
}
\arguments{
\item{collezioneValori}{collezione che contine dati della  variabile.}
}
\value{
valori modali
}
\description{
Calcola il valore della statistica moda di una variabile. Se
presenti pi&ugrave; mode, restituisce tutti i valori modali.
}
\examples{

testVal3 <-   c(1,1,2,2,1,1,2,2,17,17,17,17,17,17)
valoriModali(testVal3)

testVal4 <-   c(11,11,12,NA,11,11,12,12,12,13,14,16,15)
valoriModali(testVal4)


}
