% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineTaste.R
\docType{data}
\name{wineTaste}
\alias{wineTaste}
\title{Caso di studio wineTaste}
\format{Vedere il manuale del corso.}
\usage{
wineTaste
}
\description{
Un dataset che contiene il caso di studio wineTaste.
}
\keyword{datasets}
